Management - refactoring, extract api layer for customers module
