+++ management-service/src/api/customers/controllers/command/index.ts
@@ -1,0 +1,2 @@
+export { AddCustomerController } from './AddCustomer.controller';
+export { RemoveCustomerController } from './RemoveCustomer.controller';
