+++ management-service/src/api/customers/controllers/query/index.ts
@@ -1,0 +1,2 @@
+export { GetCustomersController } from './GetCustomers.controller';
+export { GetCustomerController } from './GetCustomer.controller';
