+++ management-service/src/api/customers/schemas/index.ts
@@ -1,0 +1,1 @@
+export { addCustomerSchema } from './AddCustomer.schema';
