+++ management-service/src/modules/customers/application/command/removeCustomer/index.ts
@@ -1,2 +1,3 @@
-export { RemoveCustomerController } from './RemoveCustomer.controller';
 export { RemoveCustomerHandler } from './RemoveCustomer.handler';
+export { RemoveCustomerCommand } from './RemoveCustomer.command';
+export { RemoveCustomerErrors } from './RemoveCustomer.errors';
