+++ management-service/src/modules/customers/application/query/getCustomer/GetCustomer.query.ts
@@ -1,0 +1,6 @@
+export class GetCustomerQuery {
+  constructor(
+    public readonly customerId: string,
+    public readonly facilityId: string,
+  ) {}
+}
