+++ management-service/src/modules/customers/application/query/getCustomer/index.ts
@@ -1,1 +1,3 @@
-export { GetCustomerController } from './GetCustomer.controller';
+export { GetCustomerErrors } from './GetCustomer.errors';
+export { GetCustomerHandler } from './GetCustomer.handler';
+export { GetCustomerQuery } from './GetCustomer.query';
