+++ management-service/src/modules/customers/application/query/getCustomers/GetCustomers.query.ts
@@ -1,0 +1,6 @@
+export class GetCustomersQuery {
+  constructor(
+    public readonly facilityId: string,
+    public readonly params: any,
+  ) {}
+}
