+++ management-service/src/modules/customers/application/query/getCustomers/index.ts
@@ -1,1 +1,3 @@
-export { GetCustomersController } from './GetCustomers.controller';
+export { GetCustomersHandler } from './GetCustomers.handler';
+export { GetCustomersErrors } from './GetCustomers.errors';
+export { GetCustomersQuery } from './GetCustomers.query';
