+++ management-service/src/modules/customers/customer.module.ts
@@ -1,19 +1,23 @@
 import { Module } from '@nestjs/common';
 import { CqrsModule } from '@nestjs/cqrs';
 
-import {
-  AddCustomerController,
-  AddCustomerHandler,
-} from './application/command/addCustomer';
-import {
-  RemoveCustomerController,
-  RemoveCustomerHandler,
-} from './application/command/removeCustomer';
-import { GetCustomerController } from './application/query/getCustomer';
-import { GetCustomersController } from './application/query/getCustomers';
 import { FacilityModule } from '../facilities/facility.module';
 import { providers } from './customer.providers';
 import { DbModule } from '../../db.module';
+
+import { AddCustomerHandler } from './application/command/addCustomer';
+import { RemoveCustomerHandler } from './application/command/removeCustomer';
+import { GetCustomerHandler } from './application/query/getCustomer';
+import { GetCustomersHandler } from './application/query/getCustomers';
+
+import {
+  AddCustomerController,
+  RemoveCustomerController,
+} from '../../api/customers/controllers/command';
+import {
+  GetCustomerController,
+  GetCustomersController,
+} from '../../api/customers/controllers/query';
 
 @Module({
   imports: [CqrsModule, DbModule, FacilityModule],
@@ -23,6 +27,12 @@ import { DbModule } from '../../db.module';
     GetCustomerController,
     GetCustomersController,
   ],
-  providers: [AddCustomerHandler, RemoveCustomerHandler, ...providers],
+  providers: [
+    AddCustomerHandler,
+    RemoveCustomerHandler,
+    GetCustomerHandler,
+    GetCustomersHandler,
+    ...providers,
+  ],
 })
 export class CustomerModule {}
